+++ tests/utils/email-value-object.util.spec.ts
@@ -44,4 +44,56 @@ describe('email-value-object.util', () => {
 		).getResult();
 		expect(valueObject.getDomain()).toBe('domain.com');
 	});
+
+	it('should create value object with success', () => {
+		const valueObject = EmailValueObject.create(
+			'username.nickname@domain.com'
+		);
+		expect(valueObject.isSuccess).toBeTruthy();
+		expect(valueObject.getResult().value).toBe(
+			'username.nickname@domain.com'
+		);
+	});
+
+	it('should create value object with success', () => {
+		const valueObject = EmailValueObject.create('rocio65@gmail.com');
+		expect(valueObject.isSuccess).toBeTruthy();
+		expect(valueObject.getResult().value).toBe('rocio65@gmail.com');
+	});
+
+	it('should create value object with success', () => {
+		const valueObject = EmailValueObject.create('rocio_65@gmail.com');
+		expect(valueObject.isSuccess).toBeTruthy();
+		expect(valueObject.getResult().value).toBe('rocio_65@gmail.com');
+	});
+
+	it('should create value object with success', () => {
+		const valueObject = EmailValueObject.create('4you@gmail.com');
+		expect(valueObject.isSuccess).toBeTruthy();
+		expect(valueObject.getResult().value).toBe('4you@gmail.com');
+	});
+
+	it('should fails if not provide a domain', () => {
+		const valueObject = EmailValueObject.create('4you@gmail');
+		expect(valueObject.isFailure).toBeTruthy();
+	});
+
+	it('should fails if provide two dots', () => {
+		const valueObject = EmailValueObject.create(
+			'my-email8950@hotmail..com'
+		);
+		expect(valueObject.isFailure).toBeTruthy();
+	});
+
+	it('should fails if provide an invalid domain', () => {
+		const valueObject = EmailValueObject.create(
+			'inVaLLiD_Email@MYDomain.longInvalid.br'
+		);
+		expect(valueObject.isFailure).toBeTruthy();
+	});
+
+	it('should fails if provide an space', () => {
+		const valueObject = EmailValueObject.create('invalid email@domain.com');
+		expect(valueObject.isFailure).toBeTruthy();
+	});
 });